refactor: fix remaining MEDIUM architecture issues
## Additional Changes

### 1. Fix Service vs Processor naming inconsistency (MEDIUM)
- Renamed `GitHubPort` struct → `GitHubService` struct in adapters/github
- GitHubPort is the interface (in domain/ports), not the implementation
- GitHubService is the adapter implementation (in adapters/github)
- Follows proper Hexagonal Architecture: Port = interface, Service = adapter

### 2. Establish model directory organization (MEDIUM)
- Moved `models/epic/epic.go` → `models/epic.go`
- Established clear rule: **Nest models only if 3+ related models OR sub-types**
- Epic had only 1 model → flattened
- Thread has 2 models → stays flat (correct)
- Story has 13+ models → nested directory (correct)

## Impact
- ✅ Build passes successfully
- ✅ Consistent naming: Port interfaces vs Service implementations
- ✅ Clear model organization criteria
